+++ test/pdm_t_octree.c
@@ -224,9 +224,9 @@ char *argv[]
 
   PDM_octree_seq_free(oct_orig);
 
-  int *weight =  malloc( n_src * sizeof(int));
+  double *weight =  malloc( n_src * sizeof(double));
   for(int i = 0; i < n_src; ++i) {
-    weight[i] = 1;
+    weight[i] = 1.;
   }
   PDM_MPI_Barrier(comm);
   // double t1 = PDM_MPI_Wtime();
